added new training + backdoor pipeline
